Extract RecordingPipeline from CameraViewController (UI modernization PR 9)
Mechanical move, no behavior/threading changes: RecordingPipeline (non-UI)
now owns the asset writer and its inputs, the six-boolean recording state
machine, the writing queue, per-frame write/crop, and saving/sending the
finished movie. The VC stays the sample-buffer delegate (captureOutput fans
out to the frame streamers, which are live preview, not recording) and keeps
the microphone-permission flow, forwarding recording frames to the pipeline.

Seams: sendMessage (actor relay for the start ack, stop response and video
resource — captures the session ref, not the VC), onRecordingStarted/Stopped
(timer overlay), onModeChanged (idle vs recording chrome), onError,
onPhotosAccessDenied.

CameraViewController: 962 -> 632 lines. New device-free tests cover asset
writer input setup including the even-rounded 4:3 crop rect and the
stop-while-idle guard. 385/385 tests green (381 + 4 new).
